Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

First of all I would propose to change your project setup.

If this will not help, create a bug describing the problem, and an example + test which reproduces the bad behavior and makes sure do not introduce a possible new regression.

Also see the comments on the patch below.

Patch-set: 1
Label: Code-Review=-1