[stagingView] Don't forget amend state after repository refreshes
The commit message and the amend state were cleared after repository
refreshes while amending. Now both is usually preserved and cleared only
if the underlying amended commit changed, e.g. because the user switched
to another branch.

We now also remember the commit message and that amend was enabled when
the Staging View or Eclipse as a whole is closed.

Bug: 454559
Change-Id: I246fd94032b8d8b0d220f2e39857bc63404e941b
Signed-off-by: Konrad Kügler <swamblumat-eclipsebugs@yahoo.de>