Update patch set 5
Patch Set 5:

Well, since initializing the scope is potentially a long-running operation, I find it better for the user to know that "something" is happening... using a null progress monitor would just do that without feedback. Do you think the pre-fetching is quick enough to bypass the progress reporting?

Patch-set: 5