Update patch set 2
Patch Set 2:

(2 comments)

Hi Markus,
I've played a bit with the code/tests. 
Sounds reasonable for me.

What would be nice: it is wrong that the changes on .gitignore file cause *full* index update: it is not necessary! Instead, simply the parents of those files has to be added as changed directories and an incremental update should be triggered on them.

Patch-set: 2