Update patch set 6
Patch Set 6:

I tried to update only sub-trees on .gitignore changes, but that does not seem to work so easily. There must be more on this somehow. I validated that the update jobs runs as expected, but the staging view does not get hold of the new state. The icon decorator for the package explorer does notice the changes in .gitignore after a second... Also one of the new unit tests is then failing when waiting for an update after deleting the .gitignore.

I'm pretty puzzled on what is going on - no idea... If you have some idea, or a little time to find out what is going on...?

Still I did the test adaptions.

Patch-set: 6
Reviewer: Gerrit User 326 <326@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0