Update patch set 1
Patch Set 1:

(3 comments)

The entire patch is actually not OK because it assumes that there is only one shell possible (and listens to this only shell) - but there can be multiple of them (multiple Eclipse windows), created and disposed etc.

To reliably answer the "isShellActive" question one has to be on SWT thread as the previous code did (surprise!) or one has to add listeners to *each* currently opened shell and another listener (or even filter) to the Display itself to get notified if shells are created or disposed.

Not sure if this is worth the effort, but the current patch isn't sufficient.

Patch-set: 1
Label: Verified=0