Update patch set 1
Patch Set 1: Code-Review-1

In this patchset the following 2 cases still use a workspace lock:

* org.eclipse.egit.ui.internal.repository.tree.command.SubmoduleUpdateCommand:
Here we need to first determine the submodules of the given repository. This itself needs to be done in the background since it's potentially long running if the git index is large. So it looks like we first need a job locking on the super-project repository to determine the submodules and then run the update commands in a second job locking on the submodule repositories

* org.eclipse.egit.ui.internal.branch.BranchOperationUI.start(): not sure if we can live without a workspace lock since BranchProjectTracker may need to re-create projects

Patch-set: 1
Reviewer: Gerrit User 4 <4@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1