[stagingview] "Updating index" should indicate that view is "busy"
If we are linked, each time IndexDiffUpdateJob starts, indicate that the
view is busy (e.g. reload() will trigger this job in background!).

This is not the best solution if one has many repositories: on Eclipse
startup the view could be shown busy even if the right repository is
already updated, but practically it works fine. 

Usually one has not really many *huge* repositories in the same Eclipse
workspace which can delay startup for long time, and after startup is
done, index refresh usually happens for the current repository only
because index diff updates are started only on repositories where user
makes some changes or selects elements.

Bug 457571
Change-Id: I5f84c70466d7cd47cc7334ce416174b857774402
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>