Scalability: don't run incremental update if there are too many changes
Always check if the number of changed files exceeds the magic limit for
incremental index updates. In case of huge repositories with lot of
ignored files (> 65000) the incremental merge of ignored lists can be
highly inefficient if the number of changed files is higher then 2000.

Added tests for job scheduling to make sure reload and update are used
as expected.

Bug 452237
Change-Id: I6cb733ae24a1eec33becb3aff0f96da151365a41
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>