Lookup active window before opening commit selection dialog
Just seen that "Compare with Previous" will fail on Eclipse 3.8 due some
bugs in HandlerUtil, if executed on a file which last version is a merge
version.

In this case we open selection dialog (which commit to use for diff),
and this makes the "active window" variable disappear. The simple fix is
to read the variable before we open a dialog.

Change-Id: I8b9cbcda45cc01aa2e8ccd65527638bb51b6e52d
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>