Update patch set 1
Patch Set 1: Code-Review-1

Adding strong dependency to JDT is a no go IMHO.
The problem is that egit can be used in Eclipse installations without JDT.

If the dependency is really needed and can't be avoided, it should be made optional and the related code should be written in the way that it can gracefully terminate if JDT classes aren't abailable. It can be done either with try/catch around related code oder with checks before if the related JDT bundle is available in the current system.

Small small issues with the patch:
 - formatting is inconsistent (check especially the spaces around "if" conditions)
 - most external class names are written full qualified in the code which avoids imports and made the code not that readable
 - commit message format: we use "Bug: number" style.

Patch-set: 1
Label: Code-Review=-1