Auto-configure Gerrit configuration if repository is cloned from Gerrit
If a repository is cloned over http or https protocol use Gerrit's
REST API endpoint "Get Version" [1] to find out if the repository is
hosted by a Gerrit server. In that case auto-configure the repository
for Gerrit. Configure the push refspec to HEAD:refs/for/master which
should be ok in most cases.

[1] https://gerrit-documentation.storage.googleapis.com/Documentation/2.11/rest-api-config.html#get-version

Bug: 460423
Change-Id: I30b916e279e0bc84f86abbffae241f8aab3fa8df
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>