Fix issue with removing repositories after path change
A repository might be moved on the file system by a user and a symbolic
link in the tree structure might be put in place to allow old references
to resolve. However, after such an action EGit fails to remove the old
path because it only looks for canonical path. The canonical path is
different after such a move.

Bug: 460118
Change-Id: I884eded8b24c2de14f0255d2371d016b4e435c28
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>