Fix behavior of EGit when repo is a symlink
The use of getCanonicalPath is problematic on POSIX file systems. If the
repo is a symlink (or one folder of the repo path) then this will be
resolved to the real file system location. 

There are multiple issues with that. Most notably is the fact that
Eclipse resource locations will not use the canonical path but an
absolute path. This translates into a couple of weird issue with
RepositoryMapping and resolving paths.

This change removes every occurence of getCanonicalPath() calls, which
should also be line with what JGit does.

Bug: 460118
Change-Id: I884eded8b24c2de14f0255d2371d016b4e435c28
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>