Commit Validation Support (post-commit) for the UI
Allows to register validators that have the chance to validate every new
commit made from the Eclipse UI. This allows to display issues in a nice
way to the user.

The drawback of this solution is that it will not work when committing
through the CLI (both JGit and CGIT).

The original intention for this feature is to bring earlier feedback to
the user. There may be checks happening on a Gerrit server that can be
done earlier. The checks need to be still on Gerrit also, to catch
people committing from the CLI.

Change-Id: I88a0fba224b1c54b145f08803063cc9495c7e582
Signed-off-by: Markus Duft <markus.duft@ssi-schaefer.com>