Wizard invoking automated import
This contributes a new Git import wizard, reusing most pages of the
current Git import wizard, but delegating the actual import actions to
the more intuitive import framework currently incubating at e4.
The benefit of this new import framework is to detect and configure
projects automatically, without user having to understand or know what
kind of project they're working on.

Change-Id: I4e3039f1b066e245fd4e26a2b93998ed2795b7ef
Signed-off-by: Mickael Istria <mistria@redhat.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>