[scalability] Use image descriptors cache for external programs
IEditorRegistry uses Program.findProgram() for file types "unknown" to
Eclipse. Unfortunately Program.findProgram() is platform dependent,
requires the caller to run in UI thread and on some platforms can be
really slow => in worst case multiple calls to retrieve image
descriptors via this API will block Eclipse UI for a long time. This
happens e.g. in the History view for commits with thousands of files or
in the Staging view with thousands of staged or unstaged files.

The proposed solution first checks if the editor is external or not, and
than uses a cache for *external* editor image descriptors, which avoids
(for files with the same extension) multiple native calls from UI thread
via IEditorRegistry.getImageDescriptor().

Bug: 464891
Change-Id: Icfbc28916656e8412ca4454046c7cee298ca2c5e
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>