Update patch set 1
Patch Set 1:

When I rename a remote, should I leave the "fetch" alone as it still contains the original remote name? All the "remote tracking" branches currently keep the original remote name too, which might be confusing to some users.
Do I have to update each branch's "remote" in the config to the new remote name?
Is there anything else I'm missing in what needs to be changed in the config/git?

Patch-set: 1