Update patch set 9
Patch Set 9:

Thank you for fixing the remaining issues!
About the locking of resources: I think additionally the "start feature" and "start hotfix" do not have to lock and thus can return null as ISchedulingRule. For "start release" it depends on whether the release will be created from a past commit, rather than HEAD.
I will fix that as soon as the CQ is done.

Patch-set: 9
Label: Verified=0