[wip] Prefer selection adaption to ResourceMapping instead of IResource
When handling the selection EGit now checks if the selected elements can
be adapted to a ResourceMapping first. Only if this is not possible it
is checked if the elements can be mapped to IResource.

This is helpful in cases where the selected element is for example a
virtual node in the project explorer. By first adapting to a
ResourceMapping the virtual node has the chance to return multiple
resources instead of only one.

Bug: 467717
Signed-off-by: Stefan Dirix <sdirix@eclipsesource.com>
Signed-off-by: Laurent Delaigue <laurent.delaigue@obeo.fr>
Signed-off-by: Mathieu Cartaud <mathieu.cartaud@obeo.fr>
Change-Id: I0c4300866ac04ae9842ff4d8d0a967c4e319ff6d