Update patch set 2
Patch Set 2:

Andrey,I was waiting for a response for "is there a better way to get the index timestamp".

It seems the fastest way would be to open the index file, parse it, and pull the stamps out manually (although the object passed into the LabelProvider does not have a method to get the hash id).

Patch-set: 2
Label: Verified=0