Do not automatically ignore paths not found by the treewalk
RepositoryUtil.isIgnored(path) is used to check if the resource should
be added to the .gitignore by IgnoreDerivedResources and
IgnoreOperation.

However if the TreeWalk can't *find* the path "isIgnored" returns
"false", but probably should return "true" in the same way as if the
mapping were missing.

To not break API, the old "isIgnored" method is deprecated and new one
"canBeAutoIgnored" is added instead, with different behavior. Added &
fixed javadoc and tests.

Bug: 470820
Change-Id: Id732684999da1561bc3812fdb2daeee7a5fae369
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>