Update patch set 2
Patch Set 2:

What happens if I have a project and I delete the subtree src/test. Previously there have been versioned resources underneath that, e.g. src/test/A.java. After this deletion I ask for isIgnored("src/test/A.java"). Since the walk we are using is looking at the working tree (and not also at Index or HEAD) this path will not be found anymore. But I wouldn't say this path "src/test/A.java" should be ignored from now on. What happens if a minute later I recreate the file "src/test/A.java". I would expect that it appears again in staging view as modified resource. If we add that path to .gitignore we wouldn't see this behaviour.

Patch-set: 2
Label: Verified=0