Test stability: another attempt to have stable connect() behavior
Refactored ConnectProviderOperation to make it easier to understand and
to be more verbose if it fails to connect. If multiple projects are to
be connected, do not fail entire task if only some of the projects have
issues. RepositoryFinder should also not silently fail during connect if
resource is unknown in workspace.

Change-Id: Ied79b1846495d7662ee40128b1a0edceb55d6860
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>