Always schedule repositories view refresh job if requested
Don't rely on Job.getState() API: "Note that job state is inherently
volatile, and in most cases clients cannot rely on the result of this
method being valid by the time the result is obtained."

The fix should not add any extra overhead but avoid unstable tests and
occasional missing refreshes of the repositories view.

Cleaned up the job scheduling logic and tracing. To fix unstable tests,
made the asyncronous execution of the UI part of the job synchronous.

Change-Id: I7ffef1b5773e4f40a0719c4a0e9f940ed34b330c
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>