Introduce "Compare mode" for Git Staging view similar to History view
A Compare Mode was introduced. Therefore a new property and a action to
change the value were created. The if the property is set to false a
double click on a file in the staging view will open it in the
workbench. To allow the user to switch the mode a action was added to
the toolbar of the view and to the view menu. Furthermore a entry to the
popup-menu was added to allow the user to compare a file with index.

Bug: 473486
Change-Id: Ifba134d4585d15dddebe16eb87ae9c2231e7cf97
Signed-off-by: Jan-Ove Weichel <ovi.weichel@gmail.com>