Add default limit for number of shown lines in history
GitHistoryPage can show diffs up to 1000 lines very fast, up to 10000
lines reasonably fast, and starts to be unusable from 20000 lines
upwards.

To avoid UI freezes, a new preference is introduced in the
Team->Git->History page, in the "Show" group: "Maximum lines to show for
a diff", with default value of 1000 lines. If this preference is set,
diff output in the History view shows only specified amount of lines.

Bug: 473013
Change-Id: Ie005c179088cf94582ce730de7415a5b52caebe1
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>