Update patch set 7
Patch Set 7:

> the generated combined commit message has a couple of problems:
 > * the meta data of the original commits should be included
 > commented out (starting with a #) in order to avoid that it ends up
 > in the squashed commit's message
 > * the commit messages of the original commits should not be
 > indented since if the user wants to edit the message he typically
 > will have to remove the leading whitespace in front of the original
 > message texts
 > * text in the commit message of the squashed commit which is
 > commented out (starting with # in first column) isn't removed from
 > the squash commit's message when the commit message edit dialog is
 > finished

So there are two commands with different formats:
- MergeCommand, which generates the format we currently see
- RebaseCommand, which generates a format that is easier to edit, and the one you would prefer

I see your point, but I would keep the current implementation for now, simply because it conforms with the reference implementation (see my comment in 471026).
I think this will make adoption of "EGit Gitflow" easier in the short term. In the long term, it would be good to have some user feedback about this.
Maybe we should request an enhancement on GitHub and see what the Gitflow community thinks about this.

Patch-set: 7
Label: Verified=0