Respect port defined in remote for Gerrit configuration.
If the push URI already is ssh, it most likely already has a port,
and if the user cloned the repo from Gerrit, it will even be correct.
Replacing that by Gerrit's default ssh port of 29418 is most confusing
if that Gerrit runs its sshd on a non-standard port.

Bug 475740
Change-Id: I6043be2c101e5efd68de66bfff851b03f0ee79b2
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>