Update patch set 1
Patch Set 1: Code-Review-1

(8 comments)

if a repository already contains a commit and a master branch and the user chooses a non-default name e.g. "master1" for the master branch this branch doesn't exist at that point. If later a release is finished gitflow tries to switch to the non existing non-default master branch "master1". Since this doesn't exist it raises an internal error:

An internal error occurred during: "Finishing release...".
No branch 'master1' found.

Would be better to ask the user on init to create the non-default master branch or create it as part of init.

Patch-set: 1
Label: Code-Review=-1