Update patch set 2
Patch Set 2: Code-Review-1

After enabling this I see now the Team-> menu with all Egit contributions on Java working set, this is nice, but... 

None of the actions shown in this menu is enabled, and none of the children drop-down menus after Team-> are functional.

Adding this would only make users angry because they would see those lots of not usable menus and ask: "what the egit developers are smoking, why all this does not work"?

I would +1 on the change if we would enable all those disabled actions (if appropriate) and *hide* all actions which are not applicable on a resource mapping.

Patch-set: 2
Label: Code-Review=-1