Fix "Import changed projects" triggered from filtered history
"Import changed projects" imported too many projects when started from
history view when it's filtered to some selected project.

Change base class of Import Changed Projects command to fix this. With
the new base class the determination of files that have been changed
with a certain commit is much easier. So we can avoid the import of
unwanted projects as reported in an Eclipse Bug.

Bug: 462977
Change-Id: Ifcfd209f33c863d93b0b7e578f83602faef2d26a
Signed-off-by: Felix Otto <felix.otto@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>