Fix EclipseAuthenticator which was breaking platform's authenticator
When EclipseAuthenticator didn't find a proxy configuration it broke
platform's NetAuthenticator which shows a password dialog. If there is
no proxy configuration delegate to NetAuthenticator to fix this.

NetAuthenticator should be enhanced to respect proxy configuration so
that we can remove EclipseAuthenticator. Until then accept dependency to
an internal platform package.

Bug: 368882
Change-Id: Ibf5c587c0a9ff549faf6ef24175e0fb986be82f5
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>