Update patch set 2
Patch Set 2:

Thanks for the review ;)

The actions only work when ALL projects in the WorkingSet are connected with the same Git repository.

We could hide the actions when this is not the case. Would you be okay with that?

As a next step the actions (at least some) could be modified, to also accept files from different repositories (for example the ReplaceWithHEAD action) and filter all files which are not in a repository.

Patch-set: 2