Update patch set 2
Patch Set 2:

> Thanks for the review ;)
 > 
 > The actions only work when ALL projects in the WorkingSet are
 > connected with the same Git repository.

This was the case, and none of the actions worked for me. I've tried both cases (same/different repos), and for different repos the Team menu was almost empty.

 > We could hide the actions when this is not the case. Would you be
 > okay with that?

Sure, this is what I would expect as user.
 
 > As a next step the actions (at least some) could be modified, to
 > also accept files from different repositories (for example the
 > ReplaceWithHEAD action) and filter all files which are not in a
 > repository.

I think this changes should go to this patch, they do not make sense without each other.

Patch-set: 2