Update patch set 3
Patch Set 3:

I added additional tests which hide all actions in the following two cases: 1) The resources map to different repositories and 2) Not all resources map to a repository.

I did not touch the Team actions since they are already enabled on ResourceMappings.

When the actions are modified to also support selections of resources pointing to different or no repositories, the individual checks can simply be removed in the plugin.xml

Patch-set: 3
Label: Verified=0