Update patch set 5
Patch Set 5:

(5 comments)

Hi Christian,

Thank you very much for the review!

I added a test extension as you requested, so that you can easily see how it looks like.
I also added a test to check that MergeOperations do use the preferred strategy.

I fixed the rest, removing the need for models to be active, and not hard-coding the default strategy.
That led me to make more changes than I imagined, but there we are.

Patch-set: 5
Label: Verified=0