Update patch set 3
Patch Set 3: -Code-Review

in the first try I once observed that the commit button in the staging view wasn't disabled despite the fact that the changes to be commited had errors and warnings and checking was on. Looks like some initialization problem though I couldn't reproduce that.
/\
maybe because the files wasn't in staged changes?

Patch-set: 3
Label: Code-Review=0