Fix hyperlink display in SpellcheckableMessageArea
This is first and foremost a syntax coloring problem, and should be
handled as such. This change implements that:

* Adds a token scanner that recognizes hyperlinks using the configured
  hyperlink detectors.
* Adds a presentation reconciler with a damager/repairer that uses this
  token scanner.
* Gets rid of the overridden hideHyperlinks() method -- if we do proper
  document partitioning, this definitely wrecks havoc.
* Uses SWT.UNDERLINE_LINK to style hyperlink partitions.
* Adds a few unit tests for the tokenizing.
* Adds org.objenesis to required bundles in org.eclipse.egit.ui.test
  to enable mocking.

In my manual UI tests, this syntax-colors hyperlinks nicely in the
staging view, spell-checking works, auto-wrap works, whitespace display
works, the "hand" pointer appears only over hyperlinks, and bug 471355
is solved. It also has no adverse effects on the commit viewer or the
git history view.

Bug: 471355
Change-Id: I623ead62d9600028973353cf63183638a1e06a3e
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>