Fix sorting of nodes in repositories view
RepositoriesViewSorter re-implemented sorting of nodes which is already
implemented in RepositoryTreeNode in a naïve way. Instead delegate the
comparison to RepositoryNode.compare().

Also improve RepositoryNode.compare() using CommonUtil comparators which
ensure that sorting is case-insensitive and sort names containing
numbers in natural order. This ensures all objects shown in the
repositories view are sorted accordingly.

This also fixes the problem that since b9dad5fa the order of repository
nodes flipped when its status became dirty/clean. This was caused by
RepositoriesViewSorter which compared the node label texts which changes
when the dirty decorator "> " is added or removed.

Change-Id: I29bfc47711ae0bd669786793ac7b40aea7fb8f55
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>