Use Apache httpclient for git http transport
This is a prerequisite to fix bug 368882: EGit registers
EclipseAuthenticator as the default JVM Authenticator to configure proxy
settings which platform's NetAuthenticator ignores. This breaks other
bundles relying on NetAuthenticator's password dialog which it raises if
a HTTP request requires authentication. Hence use Apache httpclient
which doesn't use the JVMs default authenticator to get logon
credentials.

Change-Id: Ice07b09142866fdc3d8d3cb09c8f652c401dbfe9
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>