Update patch set 7
Patch Set 7:

> this is confusing me (though I am not using git-flow so I lack
 > experience):
 > 
 > - after creating a release always master and develop point at the
 > same commit, so how can you distinguish if the tag is tagging
 > develop or master ?

This will change as soon as the other change-set (NO_FF) is merged. Then every "release finish" creates two merge commits. One on develop, one on master. The one on master should be tagged.
Probably it would have been a good idea in this one case, to fix both issues at once.

 > - I can only start a hotfix branch if develop is checked out, but
 > not if I check out master or a release tag. This contradicts my
 > understanding of http://nvie.com/posts/a-successful-git-branching-model/#finishing-a-release-branch

Yes, they should be enabled (see also 473647), but this is a minor issue at the moment, I would say. Because when working with Gitflow, you are usually on a feature branch, or on develop (unless you are the release manager).

 > which says hotfixes are meant to quickly fix a production problem ,
 > so I'd expect that hotfix branches are created from master and not
 > from develop

Yes, hotfix branches must be created from master (or possibly from a past release). If they are not, that would be a bug.

 > - the GitFlow menu in the history view doesn't show the same
 > actions enabled compared to the GitFlow menu on the same repository
 > in the repositories view, so I don't know which one is right

Someone, mentioned this in a bug report as well. Apparently it is not intuitive. I would argue that the only meaningful thing you can do in context of a commit, is creating a release, but I have no strong opinion on that.

Patch-set: 7