Update patch set 9
Patch Set 9: Code-Review-1

Changing the option "Check for warnings/errors" using the preference editor now is correctly reflected in staging view when clicking "Apply" or "Finish".

If this option is switched on I can switch it off locally in the staging view by clicking "Ignore warnings and errors". But this isn't persisted immediately and not reflected in the preference so I can reach the inconsistent state shown in http://i.imgur.com/ib38UKp.png.

As I said in an earlier comment this breaks the rule that option changes for Eclipse views should be always persisted immediately. Only editors (including preference editors) should have a dirty state. Please fix this as this non-standard behavior will likely confuse the users.

Patch-set: 9
Label: Code-Review=-1