Update patch set 9
Patch Set 9:

Is this how you want it to work?
I ask, because I want to be sure. For my understanding the preference could be set only via preference editor and the 'ignore warning and errors' checkbox appears only when the preference is set as true. It shouldn't set the preference (in my opinion), because in ideal world no file with errors/warnings should be commited, but sometimes there's an error/warning in project that has been there for ages and everybody is afraid to touch it so you're aware of that and commit with your changes without touching the rest. So normally I've got preference set to check errors/warnings and the checkbox unticked, but for 1 commit per week i'll tick the 'ignore warnings errors' checkbox, because I have to. For me it shouldn't be persisted. If you want it done your way, it's ok

Patch-set: 9