Update patch set 9
Patch Set 9:

> Changing the option "Check for warnings/errors" using the
 > preference editor now is correctly reflected in staging view when
 > clicking "Apply" or "Finish".
 > 
 > If this option is switched on I can switch it off locally in the
 > staging view by clicking "Ignore warnings and errors". But this
 > isn't persisted immediately and not reflected in the preference so
 > I can reach the inconsistent state shown in http://i.imgur.com/ib38UKp.png.
 > 
 > As I said in an earlier comment this breaks the rule that option
 > changes for Eclipse views should be always persisted immediately.
 > Only editors (including preference editors) should have a dirty
 > state. Please fix this as this non-standard behavior will likely
 > confuse the users.

I second to Pawel think this checkbox schould not persist the state but only work for the current commit.

I think I would not like this option to be persisted from the view or commit dialog, since after first use the warning will never be shown again.

Might be we can reword the message to indicate the fact that this state is not persisted, and show something like: "Mute warnings for now" or "Ignore warnings for this commit"?

Another solution would be to put a text *link* link at the end of the warning message which would open the preferences, and don't offer checkbox at all?

What I can imagine is that people would definitely love this check for errors but hate for warnings. No one commits code with compile errors bit we have lot of legacy code with warnings. So might be we also should have 2 pereferences, which user can turn on/off, and so offering one checkbox will not make everyone happy, and tho checkboxes are too much.

What about to always show the warning message but not always disable commit bittons? So instead of the checkbox show 

"Staged files contain errors or warnings. Commit not allowed. Configure..."

or 

"Staged files contain errors or warnings. Configure..."

And options would be radio-boxes, something like

"If files to commit have warnings or errors":
* "Allow commit with warnings and errors"
* "Allow commit if files have warnings"
* "Allow commit but show warning message"
* "Don't allow commit"
* "Never check for warnings or errors"

Patch-set: 9