Update patch set 9
Patch Set 9: -Code-Review

I wanted to implement such a feature, because I used to commit some files with trash like unused local variables, however I approve with Andrey, because it's more general and we do it for the users. I think that it would be also nice to be able to configure those options that Andrey wrote for particular repo. Like in general no warnings/errors are allowed, but for project X warnings are just fine, because it's legacy code with some warnings that I would not like to touch.

Patch-set: 9