Update patch set 9
Patch Set 9:

> > Might be we can reword the message to indicate the fact that this
 > > state is not persisted, and show something like: "Mute warnings
 > for
 > > now" or "Ignore warnings for this commit"?
 > 
 > now I get what you want, rewording the message would definitively
 > help to
 > avoid my misunderstanding
 
So the checkbox should be "Ignore errors/warnings for this commit"? Any other proposals?

 > I like the idea, I'd allow to configure if warning should be shown
 > if there are warnings or errors,
 > and allow to configure if this blocks commit button. I would use
 > checkboxes
 > instead of radio buttons then we don't need to spell out all
 > permutations
 > 
 > Warn if there are
 > - warnings
 > - errors
 > Block commit if there are
 > - warnings
 > - errors

I like your proposal, but let us use checkbox *and* combos in two blocks:

[x] Warn if there are: <combo with 2 choices>
 [errors]
 [warnings and errors] (default)

Next is *indented* block, disabled if the parent "Warn" checkbox is disabled:

   [x] Block commit if there are: <combo>
     [errors] (default)
     [warnings and errors]

So we end up with 4 preferences:

1) commit_show_warning: true/false, default is true

2) commit_show_warning_type: "warnings_and_errors" / "errors", default is "warnings_and_errors"

3) commit_block: true/false, default is true

4) commit_block_type: "warnings_and_errors" / "errors", default is "errors"

If the "commit_show_warning" is selected, this should automatically enable "commit_block", and if it is unchecked, it should disable "commit_block". The reason is that we don't want to *silently* block commit.

Pawel: although I agree that different projects can have different policies, I would not make those preferences project related yet, simply because this will be a little bit complicated since in one repository we can have multiple differently configured projects. We can add this later.

Pawel, you have to change the logic a little bit to differentiate:
* we can show message but do not disable the dialog
* the message can be shown for errors only or for errors and warnings
* buttons can be sdisabled for errors only or for errors and warnings

Patch-set: 9
Label: Verified=0