Obey hyperlink preferences in SpellcheckableMessageArea
and in CommitMessageViewer in git history.

1. Only open hyperlinks on direct click in read-only views; in
   editable views, this is too confusing and too often opens the
   hyperlink when the user just wants to set the caret.
2. Make the viewer react on changes in the general hyperlinking
   preferences. Both for the master switch and for individual
   contributed hyperlink detectors.
3. Don't even syntax-color hyperlinks detected by a disabled
   hyperlink detector.
4. Always keep the special git history navigation links enabled in
   the CommitMessageViewer (KnownHyperlinksDetector).
   
The JFace hyperlinking infrastructure actually only supports hyperlink
detectors that either open on a specified modifier key combination, or
on the default combination, which is the one the viewer sets, which is
normally the one from the global preferences. If a viewer thus sets
SWT.NONE to support direct hyperlinking, the global preferences will
not take effect for hyperlink detectors that specify "default", which
leads to a rather inconsistent interface, because hyperlinks are then
only active when no key is pressed, but become inactive when the
modifiers from the global preferences are pressed.

We work around this by actually duplicating hyperlink detectors when the
viewer specifies SWT.NONE. This ensures that hyperlinks can be opened
(in read-only views) either by direct click or also when the modifier
given by the global confiduration is pressed.

Bug: 411809
Change-Id: Icfcfe62b127370be36bced8b1430b764d31b75a6
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>