Update patch set 3
Patch Set 3:

> looks like now the effect of the modifier key is inverted

That's in the commit viewer in the git history, where we still use SWT.NONE as the activator. I had understood the discussion on bug 411809 such that that was the desired behavior? Read-only views should be more like web pages and open hyperlinks on direct click?

But I agree it may be confusing. We can just forego that and always obey the configured modifier key. Would be more uniform and Eclipse-like. Would be simple to change.

 > the other difference is that we always show hyperlinks whereas
 > editors only show them when moving the mouse over the link while
 > pressing the modifier, is this difference intended ?

Yes. We use syntax-coloring to highlight the places where there is a clickable hyperlink. That's what EGit always has tried to do. It also makes sense. Otherwise it's a bit obscure to figure out that you can click on some words in free text.

The maven POM editor, for instance, also syntax color hyperlinks.

Patch-set: 3
Label: Verified=0