Update patch set 3
Patch Set 3:

Did some more investigation on this apparent inversion that occurs if the viewer has
SWT.NONE as its hyperlink state mask. Apparently this is a limitation of the
hyperlinking framework of JFace. The state mask set on a viewer overrides the
default mask configured in preferences, and moreover a hyperlink detector is active
only for one mask. Two issues here:

1. HyperlinkDetectorDelegate just has fStateMask == -1 if set to "default", which
   means HyperlinkManager.isRegisteredStateMask will ignore all those individual
   settings in its loop. Effectively isRegisteredStateMask() interprets -1 as
   meaning "default as configured for the viewer", not as "default as configured
   in the global preferences".

2. HyperlinkManager.findHyperlinks(IRegion) only does something if the currently
   pressed keys:
   - if the detector's mask is -1 (default): equal to the mask configured on the
     viewer
   - otherwise: equal to the detector's mask.

Which effectively means it's impossible to have a combination where the configured
default is SWT.MOD1 and the viewer's mask is something else and have hyperlinks be
active when either MOD1 or the viewer's key is pressed: it will in this case always
only be the viewer's key. (And for hyperlink detectors that have been explicitly
configured to some non-default key, only that key.)

Therefore, I think it's really better to leave the viewer's key mask at the globally
configured default. In other words: never set SWT.NONE as key mask for our viewers, whether read-only or not.

We'll still syntax-color hyperlinks so that users see where they are; to open them, the configured key (combination) must be pressed. Unless the user configures explicitly SWT.NONE in the global preferences, no links will open on direct click. (Including the special git history navigation links for SHA1s, branches and tags in the CommitMessageViewer.)

Does that sound good?

Patch-set: 3