Update patch set 1
Patch Set 1: Code-Review-2

The tests are failing because they relied on the explicit refresh. Typical scenario is that the test calls connect on a manually created project and expects that the workspace knows about it.

The actual "production" code does not need it, so the tests must be rewritten to pass an extra flag to do the refresh after connect.

Patch-set: 1