RespositorySearchDialog: correctly detect already added repositories
In commit 2f9171f we started using the canonical path for repositories
because org.eclipse.core.internal.resources.Project uses that, so that
EGit and the Eclipse resource tree have the same idea of where projects
are. Otherwise EGit did not detect all projects belonging to a
repository.

However, this broke the detection of already imported repositories in
the RespositorySearchDialog. I had missed that because this dialog uses
always case-sensitive string comparisons for file paths, but none of
the other usual methods to compare paths (IPath, File, Path).

To fix this, we now compare both the absolute path (as before) _and_
the canonical path. This correctly catches both repositories added with
a non-canonical path as well as repositories added with a canonical
path.

The display has been left unchanged; I considered it too confusing
for the user if he entered a directory, but the list would display
potentially different canonical paths. Also resetting the directory
input to the canonical path might have been rather surprising.

As far as I see this is the only place in EGit UI where plain strings
are used as paths.

Change-Id: I8f5acf1c15b01d461a71690cdf03da528a83086b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>