Update patch set 2
Patch Set 2:

In fact, I'm not sure asking the user to open the project is the right approach at all.

In the CommitDialog, I can compare files from closed projects or files that are in no project at all easily against the index. Would be better if that just worked in the staging view, too. After all, the file exists in the git working directory.

Patch-set: 2